Added tab bar to main router
