+++ src/app/navigators/accounts.navigator.tsx
@@ -1,16 +1,29 @@
-import {createBottomTabNavigator} from '@react-navigation/bottom-tabs';
+import {createNativeStackNavigator} from '@react-navigation/native-stack';
+
 import {ACCOUNTS_NAVIGATOR_SCREEN_NAMES} from '../../constants/navigation';
 import {TransactionsByAccountScreen} from '../screens/accounts/transactions-by-account.screen';
+import {TransactionsByTagScreen} from '../screens/transactions-by-tag.screen';
+import {AccountsScreen} from '../screens/accounts.screen';
 
-const Tab = createBottomTabNavigator();
+const Stack = createNativeStackNavigator();
 
 export const AccountsNavigator = () => {
   return (
-    <Tab.Navigator>
-      <Tab.Screen
-        name={ACCOUNTS_NAVIGATOR_SCREEN_NAMES.TRANSACTIONS_BY_ACCOUNT}
+    <Stack.Navigator
+      initialRouteName={ACCOUNTS_NAVIGATOR_SCREEN_NAMES.ACCOUNTS}
+      screenOptions={{headerShown: false}}>
+      <Stack.Screen
+        name={ACCOUNTS_NAVIGATOR_SCREEN_NAMES.ACCOUNTS}
+        component={AccountsScreen}
+      />
+      <Stack.Screen
+        name={ACCOUNTS_NAVIGATOR_SCREEN_NAMES.TRANSACTIONS}
         component={TransactionsByAccountScreen}
       />
-    </Tab.Navigator>
+      <Stack.Screen
+        name={ACCOUNTS_NAVIGATOR_SCREEN_NAMES.TRANSACTIONS_BY_TAG_ID}
+        component={TransactionsByTagScreen}
+      />
+    </Stack.Navigator>
   );
 };
