+++ src/constants/navigation.ts
@@ -1,6 +1,6 @@
 export const NAVIGATOR_NAMES = {
-  ACCOUNTS: 'ACCOUNTS',
-  GROUPS: 'GROUPS',
+  ACCOUNTS: 'ACCOUNTS_NAVIGATOR',
+  GROUPS: 'GROUPS_NAVIGATOR',
 };
 
 export const ACCOUNTS_NAVIGATOR_SCREEN_NAMES = {
